+++ sktime/classification/__init__.py
@@ -2,6 +2,8 @@
 """Classifier Base."""
 __all__ = [
     "BaseClassifier",
+    "BaseEarlyClassifier",
 ]
 
 from sktime.classification.base import BaseClassifier
+from sktime.classification.early_classification_base import BaseEarlyClassifier
