--- sktime/classification/early_classification/base/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Early classification time series classifiers."""
-__all__ = [
-    "BaseEarlyClassifier",
-]
-
-from sktime.classification.early_classification.base._base import BaseEarlyClassifier
